Chore: 도메인이 맞게 api 분리
