+++ src/types/payment.ts
@@ -1,29 +1,3 @@
-type GroupJoinProps = {
-  merchantUid: string;
-  groupId: number;
-};
-
-type GroupJoinRequestData = {
-  merchantUid: string;
-};
-
-type GroupJoinResponseData = {
-  orderId: string;
-};
-
-type GroupJoinConfirmProps = {
-  merchantUid: string;
-  groupId: number;
-  impUid: string;
-  orderId: string;
-};
-
-type GroupJoinConfirmRequestData = {
-  merchantUid: string; // 주문 식별을 위한 식별자
-  impUid: string; // 결제 완료 후, 포트원으로부터 받은 값
-  orderId: string; // 주문하기 시에, 서버 측으로부터 받은 값
-};
-
 type PortOneProps = {
   clubId: number;
   pg: string;
@@ -49,12 +23,4 @@ type PortOneRequestData = {
   buyer_postcode: string;
 };
 
-export type {
-  GroupJoinProps,
-  GroupJoinRequestData,
-  GroupJoinResponseData,
-  GroupJoinConfirmProps,
-  GroupJoinConfirmRequestData,
-  PortOneProps,
-  PortOneRequestData,
-};
+export type { PortOneProps, PortOneRequestData };
